Remove PickAWinner dependency
Move SaturationLightnessPicker into Slate itself.

This solves static build issues that I can't be bothered figuring out.
PickAWinner was copied from Qt Quick Controls 1, and so it has to be
kept up-to-date with any improvements that were made since it was
copied, which it has not been.

There is no benefit to having it as a 3rd party module anyway, as it's
easier for people to just copy three files than it is to add a whole
submodule to their project.

This change also removes one more dependency to private Qt API.