Remove SplitView handle visuals
I was thinking about making a view option to lock the splitviews
to hide the handles (which make the UI too busy and don't really suit it),
but removing them altogether is easier and looks nicer. It also has
the advantage that they can be resized at any time.

This change also reduces the gap between the panels to match the new
gap between the canvas and the panel splitview.