Remove UI state compatibility code
The project file format was changed in this version, which means that certain UI
state might not load correctly if using an old project file with this version of
Slate. If you experience any issues, Slate v0.8.0 can be used to open the old
project and save it using the new format. Once this has been done, it will be
compatible with this version.