Add a swatch-based textured fill
The idea behind this feature is to allow the creation of random
fill textures based on a palette/swatch of colours that can be
hand-picked, as opposed to the existing variance (i.e. random HSL
modification) texture fill.

Currently the palette can be added to by selecting a region of
the image and selecting "Add to Textured Fill Swatch..." from the
Edit menu, which will also switch the fill type and bring up the
textured fill settings dialog. Colours can also be removed in the
dialog.

Closes #74.