Remove the requirement that notes had to be within the project bounds
This caused a weird case where it wasn't possible to drag a note to the right
or bottom edges of the project when zoomed in, which is unavoidable as long
as the note geometry stays the same at different zoom levels (which we want).

Allowing them to go out of bounds solves this problem and also makes them
more useful.